backend/requirements.txt
django == 3.1.1
djangorestframework == 3.11.1
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5

gunicorn==20.0.4

pillow == 7.2.0
psycopg2-binary == 2.8.6

django-cors-headers==3.5.0

django-filter==2.4.0
django-import-export==2.4.0

django-3-jet==1.0.8
